+++ packages/core/injected-metadata/core-injected-metadata-browser-internal/README.md
@@ -1,3 +1,3 @@
 # @kbn/core-injected-metadata-browser-internal
 
-Empty package generated by @kbn/generate
+This package contains the implementation and internal types of the browser-side injectedMedata service.
