+++ packages/core/injected-metadata/core-injected-metadata-browser-mocks/README.md
@@ -1,3 +1,3 @@
 # @kbn/core-injected-metadata-browser-mocks
 
-Empty package generated by @kbn/generate
+This package contains the public and internal mocks for the injected medata service.
