+++ packages/core/injected-metadata/core-injected-metadata-browser/README.md
@@ -1,3 +1,3 @@
 # @kbn/core-injected-metadata-browser
 
-Empty package generated by @kbn/generate
+This package contains the browser public types for the injectedMedata core service.
