+++ packages/core/injected-metadata/core-injected-metadata-browser/src/index.ts
@@ -6,6 +6,4 @@
  * Side Public License, v 1.
  */
 
-export function foo() {
-  return 'hello world';
-}
+export type { InjectedMetadataStart, InjectedMetadataSetup } from './contract';
