+++ packages/core/injected-metadata/core-injected-metadata-common-internal/README.md
@@ -1,3 +1,3 @@
 # @kbn/core-injected-metadata-common-internal
 
-Empty package generated by @kbn/generate
+This package contains the common internal types for the injectedMedata core domain.
