--- src/core/public/injected_metadata/index.ts
@@ -1,14 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0 and the Server Side Public License, v 1; you may not use this file except
- * in compliance with, at your election, the Elastic License 2.0 or the Server
- * Side Public License, v 1.
- */
-export { InjectedMetadataService } from './injected_metadata_service';
-export type {
-  InjectedMetadataParams,
-  InjectedMetadataSetup,
-  InjectedMetadataStart,
-  InjectedPluginMetadata,
-} from './injected_metadata_service';
