+++ src/core/server/plugins/types.ts
@@ -17,7 +17,7 @@ import type {
   PackageInfo,
   ConfigDeprecationProvider,
 } from '@kbn/config';
-import type { PluginName, PluginOpaqueId } from '@kbn/core-base-common';
+import type { PluginName, PluginOpaqueId, PluginType } from '@kbn/core-base-common';
 
 import { ElasticsearchConfigType } from '../elasticsearch/elasticsearch_config';
 import { SavedObjectsConfigType } from '../saved_objects/saved_objects_config';
@@ -26,7 +26,12 @@ import { CorePreboot, CoreSetup, CoreStart } from '..';
 type Maybe<T> = T | undefined;
 
 // re-exporting for now to avoid adapting all imports, will be removed later on in the migration process
-export type { PluginName, PluginOpaqueId } from '@kbn/core-base-common';
+export type {
+  PluginName,
+  PluginOpaqueId,
+  PluginType,
+  DiscoveredPlugin,
+} from '@kbn/core-base-common';
 
 /**
  * Dedicated type for plugin configuration schema.
@@ -129,18 +134,6 @@ export type MakeUsageFromSchema<T> = {
     ? MakeUsageFromSchema<T[Key]> | boolean
     : boolean;
 };
-
-/** @public */
-export enum PluginType {
-  /**
-   * Preboot plugins are special-purpose plugins that only function during preboot stage.
-   */
-  preboot = 'preboot',
-  /**
-   * Standard plugins are plugins that start to function as soon as Kibana is fully booted and are active until it shuts down.
-   */
-  standard = 'standard',
-}
 
 /** @internal */
 export interface PluginDependencies {
@@ -257,59 +250,6 @@ export interface PluginManifest {
    * A brief description of what this plugin does and any capabilities it provides.
    */
   readonly description?: string;
-
-  /**
-   * Specifies whether this plugin - and its required dependencies - will be enabled for anonymous pages (login page, status page when
-   * configured, etc.) Default is false.
-   */
-  readonly enabledOnAnonymousPages?: boolean;
-}
-
-/**
- * Small container object used to expose information about discovered plugins that may
- * or may not have been started.
- * @public
- */
-export interface DiscoveredPlugin {
-  /**
-   * Identifier of the plugin.
-   */
-  readonly id: PluginName;
-
-  /**
-   * Root configuration path used by the plugin, defaults to "id" in snake_case format.
-   */
-  readonly configPath: ConfigPath;
-
-  /**
-   * Type of the plugin, defaults to `standard`.
-   */
-  readonly type: PluginType;
-
-  /**
-   * An optional list of the other plugins that **must be** installed and enabled
-   * for this plugin to function properly.
-   */
-  readonly requiredPlugins: readonly PluginName[];
-
-  /**
-   * An optional list of the other plugins that if installed and enabled **may be**
-   * leveraged by this plugin for some additional functionality but otherwise are
-   * not required for this plugin to work properly.
-   */
-  readonly optionalPlugins: readonly PluginName[];
-
-  /**
-   * List of plugin ids that this plugin's UI code imports modules from that are
-   * not in `requiredPlugins`.
-   *
-   * @remarks
-   * The plugins listed here will be loaded in the browser, even if the plugin is
-   * disabled. Required by `@kbn/optimizer` to support cross-plugin imports.
-   * "core" and plugins already listed in `requiredPlugins` do not need to be
-   * duplicated here.
-   */
-  readonly requiredBundles: readonly PluginName[];
 
   /**
    * Specifies whether this plugin - and its required dependencies - will be enabled for anonymous pages (login page, status page when
